Extracting untar to a testutils
